business/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.12.0
django-livesync==0.5
django-mail-templated==2.6.5
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
idna==3.3
oauthlib==3.2.0
phonenumbers==8.12.50
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
urllib3==1.26.9
watchdog==2.1.8
websocket-client==1.3.2
